Almost Done little Finishing left

+++ notes.js
@@ -1,60 +1,110 @@
 const addBox = document.querySelector(".add-box"),
     popupBox = document.querySelector(".popup-box"),
+    popupTitle = document.querySelector("header p"),
     closeIcon = document.querySelector("header i"),
     titleTag = document.querySelector("input"),
     descTag = document.querySelector("textarea"),
     addBtn = document.querySelector("button");
 months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"]
 
-const notes = JSON.parse(localStorage.getItem("notes" || "[]"));
+const notes = JSON.parse(localStorage.getItem("notes") || "[]");
+let isUpdate =false, updateId;
 
 addBox.addEventListener("click", () => {
+    titleTag.focus();
+    addBtn.innerText = "Add Note";
     popupBox.classList.add("show");
+    document.querySelector("body").style.overflow = "hidden";
+    if(window.innerWidth > 660) titleTag.focus();
 })
 
 closeIcon.addEventListener("click", () => {
+    isUpdate=false;
+    titleTag.value = "";
+    descTag.value = "";
+    addBtn.innerText = "Add Note";
+    popupTitle.innerText = "Add a new Note";
     popupBox.classList.remove("show");
+    document.querySelector("body").style.overflow="auto";
 })
 
 function showNotes() {
-    notes.forEach((note) => {
+    document.querySelectorAll(".note").forEach(note => note.remove());
+    notes.forEach((note, index) => {
         let liTag = `<li class="note">
                      <div class="details">
-                         <p>${note.title}/p>
+                         <p>${note.title}</p>
                           <span>${note.description}</span>
                       </div>
                       <div class="bottom-content">
                          <span>${note.date}</span>
                           <div class="settings">
-                             <i class="uil uil-ellipsis-h"></i>
+                             <i onClick="showMenu(this)" class="uil uil-ellipsis-h"></i>
                              <ul class="menu">
-                             <li> <i class="uil uil-pen"></i>Edit </li>
-                             <li> <i class="uil uil-trash"></i>Delete</li>
+                             <li onclick="updateNote(${index},'${note.title}','${note.description}')"> <i class="uil uil-pen"></i>Edit </li>
+                             <li onclick="deleteNote(${index})"> <i class="uil uil-trash"></i>Delete</li>
                              </ul>
                          </div>
                      </div>
                    </li>`;
-                   addBox.insertAdjacentHTML("afterend", liTag);
+        addBox.insertAdjacentHTML("afterend", liTag);
     });
 }
 showNotes();
+
+function showMenu(elem) {
+    elem.parentElement.classList.add("show");
+    document.addEventListener("click", e => {
+        if (e.target.tagName != "I" || e.target != elem) {
+            elem.parentElement.classList.remove("show");
+        }
+    })
+}
+
+function updateNote(noteId, title, desc) {
+    isUpdate =true;
+    updateId=noteId;
+    addBox.click();
+    titleTag.value = title;
+    descTag.value = desc;
+    addBtn.innerText = "Update Note";
+    popupTitle.innerText = "Update  a Note";
+    
+}
+
+function deleteNote(noteId) {
+    let confirmDel=confirm("delete note?");
+    if(!confirmDel) return;
+    notes.splice(noteId, 1);
+    localStorage.setItem("notes", JSON.stringify(notes));
+    showNotes();
+}
 
 addBtn.addEventListener("click", e => {
     e.preventDefault();
     let noteTitle = titleTag.value,
         noteDesc = descTag.value;
-    let dateObj = new Date();
-    month = months[dateObj.getMonth()],
-        day = dateObj.getDate(),
-        year = dateObj.getFullYear();
 
-    let noteInfo = {
-        title: noteTitle, description: noteDesc,
-        date: `${month}, ${day}, ${year}`
+    if (noteTitle || noteDesc) {
+        let dateObj = new Date();
+        month = months[dateObj.getMonth()],
+            day = dateObj.getDate(),
+            year = dateObj.getFullYear();
+
+        let noteInfo = {
+            title: noteTitle, description: noteDesc,
+            date: `${month}, ${day}, ${year}`
+        }
+        if(!isUpdate){
+            notes.push(noteInfo);
+        }else{
+            isUpdate=false;
+            notes[updateId] =noteInfo;
+        }
+
+        // notes.push(noteInfo);
+        localStorage.setItem("notes", JSON.stringify(notes));
+        showNotes();
+        closeIcon.click();
     }
-
-    notes.push(noteInfo);
-    localStorage.setItem("notes", JSON.stringify(notes));
-    closeIcon.click();
-
 });
